refactor: removed old interaction logic
