+++ src/events/selectMenus/manageVotecount/replacePlayer.ts
@@ -1,6 +1,6 @@
 import { type UserSelectMenuInteraction, type CacheType } from 'discord.js';
 import { UserSelectMenu } from '../../../structures/interactions/UserSelectMenu';
-import { InteractionError } from '../../../structures/interactions';
+import { InteractionError } from '../../../structures/interactions/_Interaction';
 import { prisma } from '../../..';
 import { getOrCreateUser } from '@models/users';
 import { getVoteCounter, type FullPlayer, getPlayer } from '@models/votecounter';
