+++ src/structures/interactions/_Interaction.ts
@@ -1,6 +1,7 @@
-import { type Interaction as CoreInteraction, type MessagePayload, type BaseInteraction } from 'discord.js';
+import { type BaseInteraction } from 'discord.js';
 import fs from 'fs';
 import path from 'path';
+
 type CustomID = string;
 
 export class InteractionError extends Error {
@@ -12,6 +13,7 @@ export class InteractionError extends Error {
 
 export class Interaction {
 	private customId: CustomID;
+
 	constructor(customId: CustomID) {
 		this.customId = customId;
 		console.log(`Loaded [${this.customId}]`);
@@ -33,41 +35,6 @@ export class Interaction {
 		if (!customId) return [customIdString, undefined];
 		if (!data) return [customId, ''];
 		return [customId, data];
-	}
-
-	static async loadInteractions(newPath: string, recursive: boolean = true) {
-		const commandPath = newPath;
-
-		const loadFiles = async (dirPath: string) => {
-			try {
-				const files = fs.readdirSync(dirPath);
-				for (const file of files) {
-					const filePath = path.join(dirPath, file);
-					const stats = fs.statSync(filePath);
-					if (stats.isDirectory() && recursive) {
-						await loadFiles(filePath); // Recursive call for subdirectories
-					} else if (stats.isFile() && (file.endsWith('.ts') || file.endsWith('.js'))) {
-						try {
-							// eslint-disable-next-line @typescript-eslint/no-var-requires
-							const fileData = require(filePath);
-							for (const key in fileData) {
-								const value = fileData[key];
-								if (value.prototype instanceof Interaction) {
-									new value();
-								}
-							}
-						} catch (err) {
-							console.log(`\x1B[31mFailed to load file: \x1B[34m${file}\x1B[0m`);
-							console.error(err);
-						}
-					}
-				}
-			} catch (err) {
-				console.log(`\x1B[31mFailed to load directory: \x1B[34m${newPath}\x1B[0m`);
-			}
-		};
-
-		await loadFiles(commandPath);
 	}
 
 	async onError(i: BaseInteraction, err: unknown) {
@@ -110,8 +77,37 @@ export class Interaction {
 	}
 }
 
-export async function safeReply(i: CoreInteraction, data: string | MessagePayload) {
-	if (!i.isRepliable()) return;
-	if (i.replied) return await i.followUp(data);
-	else return await i.reply(data);
+export async function loadInteractions(newPath: string, recursive: boolean = true) {
+	const commandPath = newPath;
+
+	const loadFiles = async (dirPath: string) => {
+		try {
+			const files = fs.readdirSync(dirPath);
+			for (const file of files) {
+				const filePath = path.join(dirPath, file);
+				const stats = fs.statSync(filePath);
+				if (stats.isDirectory() && recursive) {
+					await loadFiles(filePath); // Recursive call for subdirectories
+				} else if (stats.isFile() && (file.endsWith('.ts') || file.endsWith('.js'))) {
+					try {
+						// eslint-disable-next-line @typescript-eslint/no-var-requires
+						const fileData = require(filePath);
+						for (const key in fileData) {
+							const value = fileData[key];
+							if (value.prototype instanceof Interaction) {
+								new value();
+							}
+						}
+					} catch (err) {
+						console.log(`\x1B[31mFailed to load file: \x1B[34m${file}\x1B[0m`);
+						console.error(err);
+					}
+				}
+			}
+		} catch (err) {
+			console.log(`\x1B[31mFailed to load directory: \x1B[34m${newPath}\x1B[0m`);
+		}
+	};
+
+	await loadFiles(commandPath);
 }
